# packages/contracts/openapi/control-api.yaml
openapi: 3.0.3
info:
  title: KofClub Control API
  version: 0.1.0
  description: |
    Control-plane HTTP API (accounts, clubs, roles, table configuration,
    virtual-chip ledger administration, hand history).

    Conventions (see docs/adr/ADR-009-wire-conventions.md):
    - All endpoints are versioned under `/v1` (health probes are unversioned).
    - JSON bodies use camelCase; enums are UPPERCASE strings.
    - Identifiers are UUID strings (UUIDv7 preferred).
    - Timestamps are UTC RFC 3339 strings.
    - Chip amounts are integers (int64). Floating point is never used.
    - Every response carries an `X-Request-Id` header; clients may send one.
    - State-changing requests that may be retried accept an `Idempotency-Key` header.
    - Errors always use the `ErrorEnvelope` shape with a machine-readable `code`.

    Virtual chips have no monetary value. This API intentionally contains no
    deposit, withdrawal, cash-out or payment functionality (ADR-006).
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /v1/clubs:
    post:
      tags: [clubs]
      operationId: createClub
      summary: Create a club (caller becomes OWNER)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClubRequest'
      responses:
        '201':
          description: Club created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          $ref: '#/components/responses/Error'
    get:
      tags: [clubs]
      operationId: listMyClubs
      summary: Clubs I am an active member of
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Clubs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubList'
...
  /v1/clubs/{clubId}:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    get:
      tags: [clubs]
      operationId: getClub
      summary: Club details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      tags: [clubs]
      operationId: updateClub
      summary: Rename the club or edit its description (OWNER)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClubRequest'
      responses:
        '200':
          description: Updated club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
...
  /v1/clubs/{clubId}/audit-log:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    get:
      tags: [clubs]
      operationId: listClubAuditLog
      summary: Privileged-action audit log (ADMIN+, platform admins)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of audit records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPage'
        '403':
          $ref: '#/components/responses/Error'
...
  /v1/admin/overview:
    get:
      tags: [admin]
      operationId: adminOverview
      summary: Platform operational overview (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformOverview'
        '403':
          $ref: '#/components/responses/Error'
  /v1/admin/users:
    get:
      tags: [admin]
      operationId: adminSearchUsers
      summary: Search accounts (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive prefix (username/email for users, name for clubs).
          schema:
            type: string
            maxLength: 64
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserPage'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /v1/admin/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    patch:
      tags: [admin]
      operationId: adminSetUserStatus
      summary: Suspend or reinstate an account; suspension revokes every session (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/clubs:
    get:
      tags: [admin]
      operationId: adminSearchClubs
      summary: Search clubs (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive prefix (username/email for users, name for clubs).
          schema:
            type: string
            maxLength: 64
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminClubPage'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /v1/admin/clubs/{clubId}:
    parameters:
      - $ref: '#/components/parameters/ClubId'
    patch:
      tags: [admin]
      operationId: adminSetClubStatus
      summary: Suspend (view-only) or reinstate a club (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminClub'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /v1/admin/audit-log:
    get:
      tags: [admin]
      operationId: adminAuditLog
      summary: Platform-wide audit log (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: actorUserId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: action
          in: query
          required: false
          schema:
            type: string
            pattern: '^[A-Z_]{3,64}$'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPage'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /v1/admin/risk-events:
    get:
      tags: [admin]
      operationId: adminRiskEvents
      summary: Risk cases awaiting (OPEN) or after (REVIEWED) review (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, REVIEWED]
            default: OPEN
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskEventPage'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /v1/admin/risk-events/{eventId}:
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    patch:
      tags: [admin]
      operationId: adminReviewRiskEvent
      summary: Record a review disposition (evidence stays immutable) (PLATFORM_ADMIN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRiskEventRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskEvent'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ClubId:
      name: clubId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from a previous page's `nextCursor`.
      schema:
        type: string
        maxLength: 512
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Unique key making a retried state-changing request a no-op.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    Error:
      description: Machine-readable error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
      description: UTC RFC 3339 timestamp
...
    PlatformRole:
      type: string
      enum: [USER, PLATFORM_ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, DELETED]
...
    ClubStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, CLOSED]
...
    Club:
      type: object
      additionalProperties: false
      required: [id, name, description, status, ownerUserId, createdAt, myRole]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ClubStatus'
        ownerUserId:
          $ref: '#/components/schemas/Uuid'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        myRole:
          type: string
          enum: [OWNER, ADMIN, AGENT, MEMBER, null]
          nullable: true
          description: Null when viewed through platform-admin oversight.
        memberCount:
          type: integer
        joinCode:
          type: string
          description: Visible to AGENT and above only.
    ClubList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Club'
    CreateClubRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 64
        description:
          type: string
          maxLength: 500
    UpdateClubRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 64
        description:
          type: string
          nullable: true
          maxLength: 500
...
    AuditRecord:
      type: object
      additionalProperties: false
      required:
        [
          id,
          actorUserId,
          actorUsername,
          clubId,
          action,
          objectType,
          objectId,
          before,
          after,
          requestId,
          createdAt,
        ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        actorUserId:
          type: string
          format: uuid
          nullable: true
        actorUsername:
          type: string
          nullable: true
        clubId:
          type: string
          format: uuid
          nullable: true
        action:
          type: string
        objectType:
          type: string
        objectId:
          type: string
          nullable: true
        before:
          type: object
          additionalProperties: true
          nullable: true
        after:
          type: object
          additionalProperties: true
          nullable: true
        requestId:
          type: string
          nullable: true
        createdAt:
          $ref: '#/components/schemas/Timestamp'
    AuditPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditRecord'
        nextCursor:
          type: string
          nullable: true
...
    PlatformOverview:
      type: object
      additionalProperties: false
      required: [users, clubs, tables, hands, sessions, risk, ledger, generatedAt]
      properties:
        users:
          type: object
          additionalProperties: false
          required: [total, suspended, platformAdmins]
          properties:
            total: { type: integer }
            suspended: { type: integer }
            platformAdmins: { type: integer }
        clubs:
          type: object
          additionalProperties: false
          required: [total, active, suspended]
          properties:
            total: { type: integer }
            active: { type: integer }
            suspended: { type: integer }
        tables:
          type: object
          additionalProperties: false
          required: [open, seatedPlayers]
          properties:
            open: { type: integer }
            seatedPlayers: { type: integer }
        hands:
          type: object
          additionalProperties: false
          required: [inProgress, completedLast24h, voidedLast24h]
          properties:
            inProgress: { type: integer }
            completedLast24h: { type: integer }
            voidedLast24h: { type: integer }
        sessions:
          type: object
          additionalProperties: false
          required: [active]
          properties:
            active: { type: integer }
        risk:
          type: object
          additionalProperties: false
          required: [openEvents, highSeverityOpen]
          properties:
            openEvents: { type: integer }
            highSeverityOpen: { type: integer }
        ledger:
          type: object
          additionalProperties: false
          required: [invariantViolations]
          properties:
            invariantViolations:
              type: integer
              description: Must always be 0; anything else pages the on-call (docs/runbooks).
        generatedAt:
          $ref: '#/components/schemas/Timestamp'
    AdminUser:
      type: object
      additionalProperties: false
      required: [id, username, email, status, platformRole, createdAt, clubCount, activeSessions]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        username:
          type: string
        email:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        platformRole:
          $ref: '#/components/schemas/PlatformRole'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        clubCount:
          type: integer
        activeSessions:
          type: integer
    AdminUserPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        nextCursor:
          type: string
          nullable: true
    AdminClub:
      type: object
      additionalProperties: false
      required: [id, name, status, ownerUserId, ownerUsername, memberCount, openTables, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        status:
          $ref: '#/components/schemas/ClubStatus'
        ownerUserId:
          $ref: '#/components/schemas/Uuid'
        ownerUsername:
          type: string
        memberCount:
          type: integer
        openTables:
          type: integer
        createdAt:
          $ref: '#/components/schemas/Timestamp'
    AdminClubPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminClub'
        nextCursor:
          type: string
          nullable: true
    UpdateAccountStatusRequest:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        reason:
          type: string
          minLength: 3
          maxLength: 500
          description: Recorded in the audit log.
    RiskEvent:
      type: object
      additionalProperties: false
      required:
        [
          id,
          subjectUserId,
          subjectUsername,
          clubId,
          type,
          severity,
          score,
          featureValues,
          evidenceRefs,
          createdAt,
          reviewedAt,
          reviewedBy,
          reviewedByUsername,
          disposition,
          reviewNote,
        ]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        subjectUserId:
          type: string
          format: uuid
          nullable: true
        subjectUsername:
          type: string
          nullable: true
        clubId:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        score:
          type: integer
        featureValues:
          type: object
          additionalProperties: true
        evidenceRefs:
          type: array
          items: {}
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reviewedBy:
          type: string
          format: uuid
          nullable: true
        reviewedByUsername:
          type: string
          nullable: true
        disposition:
          type: string
          enum: [DISMISSED, CONFIRMED, ESCALATED, null]
          nullable: true
        reviewNote:
          type: string
          nullable: true
    RiskEventPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RiskEvent'
        nextCursor:
          type: string
          nullable: true
    ReviewRiskEventRequest:
      type: object
      additionalProperties: false
      required: [disposition, note]
      properties:
        disposition:
          type: string
          enum: [DISMISSED, CONFIRMED, ESCALATED]
        note:
          type: string
          minLength: 3
          maxLength: 1000
...
    ErrorCode:
      type: string
      description: |
        Stable machine-readable error codes shared by HTTP and realtime APIs.
        Clients must branch on `code`, never on `message`.
      enum:
        # generic
        - VALIDATION_FAILED
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL
        - SERVICE_UNAVAILABLE
        - IDEMPOTENCY_CONFLICT
        # auth
        - AUTH_REQUIRED
        - AUTH_INVALID_CREDENTIALS
        - AUTH_TOKEN_INVALID
        - AUTH_TOKEN_EXPIRED
        - AUTH_REFRESH_INVALID
        - AUTH_REFRESH_REUSED
        - AUTH_SESSION_REVOKED
        - ACCOUNT_SUSPENDED
        - EMAIL_TAKEN
        - USERNAME_TAKEN
        - FORBIDDEN
        # clubs
        - CLUB_NOT_FOUND
        - NOT_CLUB_MEMBER
        - CLUB_BANNED
        - ALREADY_CLUB_MEMBER
        - INVITE_INVALID
        - ROLE_CHANGE_NOT_ALLOWED
        # ledger
        - INSUFFICIENT_CHIPS
        - LEDGER_INVARIANT_VIOLATION
        # tables / game
        - TABLE_NOT_FOUND
        - TABLE_CLOSED
        - TABLE_FULL
        - SEAT_TAKEN
        - ALREADY_SEATED
        - PLAYER_NOT_SEATED
        - INVALID_BUY_IN
        - NOT_YOUR_TURN
        - ILLEGAL_ACTION
        - INVALID_RAISE
        - HAND_NOT_ACTIVE
        - STALE_GAME_STATE
        - ACTION_ALREADY_PROCESSED
        - TABLE_UNAVAILABLE
        # history
        - HAND_NOT_FOUND
        # tournaments (phase 3)
        - TOURNAMENT_NOT_OPEN
    ErrorBody:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable summary (not for programmatic use).
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'